refactor: using consistent logic for getting pending tx count
